chore: Update dependencies in Cargo files
This commit adds new dependencies to the project, including `embed-resource` version 2.5.1, `vswhom` version 0.1.0, `vswhom-sys` version 0.1.3, and `winreg` version 0.52.0, along with their respective dependencies. The `Cargo.toml` file is also updated to reflect the new version of `embed-resource` as a build dependency. These changes ensure the project is up-to-date with the latest libraries for improved functionality.